+++ src/podcastfuncs.h
@@ -21,6 +21,7 @@ size_t getPodcastList(struct List *, int);
 void showPodcastInfo(struct Node *);
 void showPodcastEpisodeInfo(struct Node *);
 void fillPodcastList(struct filters);
+void fillPodcastTrendingList(struct filters);
 void playPodcast(struct Node *);
 
 #endif